Deal with systemd-user-sessions

+++ ee/server/service/embedded_scripts/linux_unlock.sh
@@ -21,4 +21,11 @@ do
 done
 
 # Remove the pam_nologin file
-rm /etc/nologin
+[ -f /etc/nologin ] && rm /etc/nologin
+
+# Enable systemd-user-sessions, a service that deletes /etc/nologin
+if [ -f /usr/lib/systemd/system/systemd-user-sessions.service ]; then
+    sed -i '/^Exec/s/=true #/=/' /usr/lib/systemd/system/systemd-user-sessions.service
+    systemctl daemon-reload
+    /usr/lib/systemd/systemd-user-sessions start
+fi
